refactor: refactoring modals + restructuring packages
