+++ src/scripts/process/background.js
@@ -1,8 +1,8 @@
-import {handleIconClick} from "./events/handleIconClick.bg";
+import {handleIconClick} from "./events/icon/iconClickListener.bg";
 // import {exportAllThreadsListener} from "../events/exportAll.bg";
-import {handleIconImage} from "./events/handleIconImage.bg";
+import {handleIconImage} from "./events/icon/iconImageListeners.bg";
 import {setLifecycleActions} from "./tasks/setLifecycleActions.bg";
-import {setClickCountListener} from "./events/clickCountListener.bg";
+import {setClickCountListener} from "./events/icon/clickCountListener.bg";
 
 background();
