+++ src/scripts/process/events/icon/clickCountListener.bg.js
@@ -1,4 +1,4 @@
-import {initClickIconCount} from "../../units/page/modals/clickCount.all";
+import {initClickIconCount} from "../../storage/clickIconCountContext.tab";
 
 export function setClickCountListener() {
   chrome.runtime.onInstalled.addListener(
