+++ src/scripts/process/events/icon/iconClickListener.bg.js
