+++ src/scripts/process/events/pageIntegrationListener.tab.js
@@ -1,4 +1,4 @@
-import {launchModalIntegration} from "../tasks/launchModalIntegration.tab";
+import {launchModalIntegration} from "../tasks/integrateModals.tab";
 
 export async function pageIntegrationListener(domain) {
   window.addEventListener('load', async function () {
