+++ src/scripts/process/tasks/launchScraperOnPage.bg.js
@@ -1,4 +1,4 @@
-import {launchScrapping} from "./launchScrapping.tab";
+import {launchScrapping} from "./scrapping.tab";
 import {clickOnListElt} from "../../units/page/interact-DOM/interact.tab";
 
 export function launchScrappingActionOnPage(request, domain, sendResponse) {
